+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sensors/camera/camera_cfg.py
@@ -12,25 +12,12 @@ from typing import List, Tuple
 # omni-isaac-orbit
 from omni.isaac.orbit.utils import configclass
 
+__all__ = ["PinholeCameraCfg", "FisheyeCameraCfg"]
+
 
 @configclass
 class PinholeCameraCfg:
-    """Configuration for a pinhole camera sensor.
-
-    The following sensor types are supported:
-
-        - "rgb": A rendered color image.
-        - "distance_to_camera": An image containing the distance to camera optical center.
-        - "distance_to_image_plane": An image containing distances of 3D points from camera plane along camera's z-axis.
-        - "normals": An image containing the local surface normal vectors at each pixel.
-        - "motion_vectors": An image containing the motion vector data at each pixel.
-        - "instance_segmentation": The instance segmentation data.
-        - "semantic_segmentation": The semantic segmentation data.
-        - "bounding_box_2d_tight": The tight 2D bounding box data (only contains non-occluded regions).
-        - "bounding_box_2d_loose": The loose 2D bounding box data (contains occluded regions).
-        - "bounding_box_3d": The 3D view space bounding box data.
-        - "occlusion": The occlusion information (such as instance id, semantic id and occluded ratio).
-    """
+    """Configuration for a pinhole camera sensor."""
 
     sensor_tick: float = 0.0
     """Simulation seconds between sensor buffers. Defaults to 0.0."""
@@ -43,7 +30,7 @@ class PinholeCameraCfg:
     semantic_types: List[str] = ["class"]
     """List of allowed semantic types the types. Defaults to ["class"].
 
-    For example, if `semantic_types` is [“class”], only the bounding boxes for prims with semantics of
+    For example, if semantic types is [“class”], only the bounding boxes for prims with semantics of
     type “class” will be retrieved.
 
     More information available at:
@@ -52,12 +39,11 @@ class PinholeCameraCfg:
 
     @configclass
     class UsdCameraCfg:
-        """USD related configuration for the sensor.
+        """USD related configuration of the sensor.
 
-        Note:
-            The parameter is kept default from USD if it is set to :obj:`None`. This includes the default
-            parameters (in case the sensor is created) or the ones set by the user (in case the sensor is
-            loaded from existing USD stage).
+        The parameter is kept default from USD if it is set to :obj:`None`. This includes the default
+        parameters (in case the sensor is created) or the ones set by the user (in case the sensor is
+        loaded from existing USD stage).
 
         Reference:
             * https://docs.omniverse.nvidia.com/prod_extensions/prod_extensions/ext_replicator/annotators_details.html
@@ -91,37 +77,10 @@ class PinholeCameraCfg:
 
 @configclass
 class FisheyeCameraCfg(PinholeCameraCfg):
-    """Configuration for a fisheye camera sensor.
-
-    The following sensor types are supported:
-
-        - "rgb": A rendered color image.
-        - "distance_to_camera": An image containing the distance to camera optical center.
-        - "distance_to_image_plane": An image containing distances of 3D points from camera plane along camera's z-axis.
-        - "normals": An image containing the local surface normal vectors at each pixel.
-        - "motion_vectors": An image containing the motion vector data at each pixel.
-        - "instance_segmentation": The instance segmentation data.
-        - "semantic_segmentation": The semantic segmentation data.
-        - "bounding_box_2d_tight": The tight 2D bounding box data (only contains non-occluded regions).
-        - "bounding_box_2d_loose": The loose 2D bounding box data (contains occluded regions).
-        - "bounding_box_3d": The 3D view space bounding box data.
-        - "occlusion": The occlusion information (such as instance id, semantic id and occluded ratio).
-
-    """
+    """Configuration for a fisheye camera sensor."""
 
     @configclass
     class UsdCameraCfg(PinholeCameraCfg.UsdCameraCfg):
-        """USD related configuration for the sensor.
-
-        Note:
-            The parameter is kept default from USD if it is set to :obj:`None`. This includes the default
-            parameters (in case the sensor is created) or the ones set by the user (in case the sensor is
-            loaded from existing USD stage).
-
-        Reference:
-            * https://docs.omniverse.nvidia.com/app_isaacsim/app_isaacsim/manual_replicator_composer_parameter_list.html#camera-parameters
-            * https://graphics.pixar.com/usd/docs/api/class_usd_geom_camera.html
-        """
 
         fisheye_nominal_width: float = None
         """Nominal width of fisheye lens model."""