+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/utils/__init__.py
@@ -14,18 +14,24 @@ Sub-module containing utilities for the Orbit framework.
 * `timer`: Provides a timer class (uses contextlib) for benchmarking.
 """
 
+from .array import TENSOR_TYPE_CONVERSIONS, TENSOR_TYPES, convert_to_torch
 from .configclass import configclass
-from .dict import class_to_dict, print_dict, update_class_from_dict, update_dict
+from .dict import class_to_dict, convert_dict_to_backend, print_dict, update_class_from_dict, update_dict
 from .string import to_camel_case, to_snake_case
 from .timer import Timer
 
 __all__ = [
+    # arrays
+    "TENSOR_TYPES",
+    "TENSOR_TYPE_CONVERSIONS",
+    "convert_to_torch",
     # config wrapper
     "configclass",
     # dictionary utilities
     "class_to_dict",
-    "update_class_from_dict",
+    "convert_dict_to_backend",
     "print_dict",
+    "update_class_from_dict",
     "update_dict",
     # string utilities
     "to_camel_case",